fix: change the placement of the showing row data function & update way of getting  description
